Added Scratch for acceleration and braking physics
Created function accelbalance that governs the acceleration physics of
the car given a car and tire object, torque on the axle, and the initial
velocity of the vehicle. This function is used as a argument into
sp.integrate.quad which calculates the speed reached after a specific
amount of time (it integrates from time=0 to a specified time).

brakebalance function governs the physics of  braking. This is left as a
second order ODE which is solved by sp.integrate.odeint.